improving code consistency for combat stat and tuned the do_consider() command

+++ vme/src/experience.cpp
@@ -383,26 +383,37 @@ int melee_bonus(class unit_data *att, class unit_data *def,
                 int hit_loc,
                 int *pAtt_weapon_type, class unit_data **pAtt_weapon,
                 int *pDef_armour_type, class unit_data **pDef_armour,
-                int primary)
+                int primary, std::string *pStat)
 {
    int att_bonus, att_wpn_knowledge;
-   int def_dex, def_bonus, def_wpn_knowledge;
+   int def_bonus, def_wpn_knowledge;
+   char buf[MAX_STRING_LENGTH];
 
    class unit_data *att_wpn;
    int att_wpn_type;
 
-   int def_armour_type;
+   int def_armour_type, tmp;
    class unit_data *def_armour;
 
    int hm;
 
-   def_dex = effective_dex(def);
+   if (pStat)
+   {
+      sprintf(buf, "%s attacking %s<br/>", UNIT_NAME(att), UNIT_NAME(def));
+      pStat->append(buf);
+      pStat->append("<u>Melee bonus explanation</u>:<br/><pre>");
+      pStat->append("                        ATT     DEF<br/>");
+   }
 
    att_bonus = CHAR_OFFENSIVE(att);
    def_bonus = CHAR_DEFENSIVE(def);
+   if (pStat)
+   {
+      sprintf(buf, "Off / Def bonus     :  %4d   %4d<br/>", att_bonus, def_bonus);
+      pStat->append(buf);
+   }
 
-   if (pAtt_weapon_type &&
-       is_in(*pAtt_weapon_type, WPN_GROUP_MAX, WPN_TREE_MAX - 1))
+   if (pAtt_weapon_type && is_in(*pAtt_weapon_type, WPN_GROUP_MAX, WPN_TREE_MAX - 1))
    {
       att_wpn = NULL;
       att_wpn_type = *pAtt_weapon_type;
@@ -419,13 +430,32 @@ int melee_bonus(class unit_data *att, class unit_data *def,
       {
          att_wpn_type = OBJ_VALUE(att_wpn, 0); /* [0] is category */
          att_wpn_knowledge = weapon_attack_skill(att, att_wpn_type);
+         if (pStat)
+         {
+            sprintf(buf, "Att Weapon          :  %s       <br/>", UNIT_NAME(att_wpn));
+            pStat->append(buf);
+         }
          if (is_in(OBJ_VALUE(att_wpn, 1), -25, 25))
-            att_bonus += OBJ_VALUE(att_wpn, 1) + OBJ_VALUE(att_wpn, 2);
+         {
+            tmp = OBJ_VALUE(att_wpn, 1) + OBJ_VALUE(att_wpn, 2);
+            att_bonus += tmp;
+            if (pStat)
+            {
+               sprintf(buf, "Att Weapon mat+mag  :  %4d       <br/>", tmp);
+               pStat->append(buf);
+            }
+         }
       }
       else
       {
          att_wpn_type = CHAR_ATTACK_TYPE(att);
          att_wpn_knowledge = weapon_attack_skill(att, att_wpn_type);
+
+         if (pStat)
+         {
+            sprintf(buf, "Att Weapon          :  None       <br/>");
+            pStat->append(buf);
+         }
       }
    }
 
@@ -439,191 +469,82 @@ int melee_bonus(class unit_data *att, class unit_data *def,
          dual_skill = CHAR_DEX(att);
 
       if (primary)
-         att_bonus -= MAX(0, 25 - (dual_skill / 4));
+         tmp = -MAX(0, 25 - (dual_skill / 4));
       else
-         att_bonus -= MAX(0, 50 - (dual_skill / 4));
+         tmp = -MAX(0, 50 - (dual_skill / 4));
+
+      att_bonus += tmp;
+
+      if (pStat)
+      {
+         sprintf(buf, "Dual wield          :  %4d       <br/>", tmp);
+         pStat->append(buf);
+      }
    }
 
    def_wpn_knowledge = weapon_defense_skill(def, att_wpn_type);
+   if (pStat)
+   {
+      sprintf(buf, "Weapon skill        :  %4d   %4d<br/>", att_wpn_knowledge, def_wpn_knowledge);
+      pStat->append(buf);
+   }
 
    if (CHAR_FIGHTING(def) != att)
+   {
       def_bonus -= 25;
 
-   /* If attacker can't see the defender, then the defender has a   */
-   /* much better effective dexterity (since attacker can't see him */
-   if (!CHAR_CAN_SEE(att, def))
-      def_bonus += 25;
-
-   if (!CHAR_CAN_SEE(def, att))
-      att_bonus += 25;
-
-   /* Slaying Weapons */
-   if (att_wpn && OBJ_VALUE(att_wpn, 3) == CHAR_RACE(def))
-      att_bonus += 25;
-
-   if ((UNIT_IS_GOOD(att) && affected_by_spell(def, ID_PROT_GOOD)) ||
-       (UNIT_IS_EVIL(att) && affected_by_spell(def, ID_PROT_EVIL)))
-      def_bonus += 20;
-
-   if ((def_armour = equipment_type(def, hit_loc, ITEM_ARMOR)))
-   {
-      def_armour_type = OBJ_VALUE(def_armour, 0);
-      if (is_in(OBJ_VALUE(def_armour, 1), -25, 25))
-         def_bonus += OBJ_VALUE(def_armour, 1) + OBJ_VALUE(def_armour, 2);
-      else
-         slog(LOG_ALL, 0, "Illegal armour bonus.");
-   }
-   else
-      def_armour_type = CHAR_NATURAL_ARMOUR(def);
-
-   if (pAtt_weapon_type)
-      *pAtt_weapon_type = att_wpn_type;
-
-   if (pAtt_weapon)
-      *pAtt_weapon = att_wpn;
-
-   if (pDef_armour_type)
-      *pDef_armour_type = def_armour_type;
-
-   if (pDef_armour)
-      *pDef_armour = def_armour;
-
-   int att_abil = weapon_attack_ability(att, att_wpn_type);
-
-   if (CHAR_AWAKE(def))
-      hm = 2* ((att_abil - def_dex) + (att_wpn_knowledge - def_wpn_knowledge/3) + 
-                att_bonus - def_bonus + CHAR_LEVEL(att) - CHAR_LEVEL(def));
-   else
-      hm = 2* (att_abil + att_wpn_knowledge + att_bonus + CHAR_LEVEL(att) - CHAR_LEVEL(def));
-
-   return MAX(-50, hm);
-}
-
-
-// Appends an explanation of melee_bonus to the string str
-//
-void stat_melee_bonus(string &str,
-                      class unit_data *att, class unit_data *def,
-                      int hit_loc,
-                      int *pAtt_weapon_type, class unit_data **pAtt_weapon,
-                      int *pDef_armour_type, class unit_data **pDef_armour,
-                      int primary)
-{
-   int att_bonus, att_wpn_knowledge;
-   int def_dex, def_bonus, def_wpn_knowledge;
-   char buf[MAX_STRING_LENGTH];
-
-   class unit_data *att_wpn;
-   int att_wpn_type;
-
-   int def_armour_type;
-   class unit_data *def_armour;
-
-   int hm, tmp;
-
-   str.append("<u>Melee bonus explanation</u>:<br/><pre>");
-   sprintf(buf, "                        ATT     DEF<br/>");
-   str.append(buf);
-
-   att_bonus = CHAR_OFFENSIVE(att);
-   def_bonus = CHAR_DEFENSIVE(def);
-   sprintf(buf, "Off / Def bonus     :  %4d   %4d<br/>", att_bonus, def_bonus);
-   str.append(buf);
-
-   tmp = 0;
-
-   if (primary)
-      att_wpn = equipment_type(att, WEAR_WIELD, ITEM_WEAPON);
-   else
-      att_wpn = equipment_type(att, WEAR_HOLD, ITEM_WEAPON);
-
-   if (att_wpn)
-   {
-      att_wpn_type = OBJ_VALUE(att_wpn, 0); /* [0] is category ??? */
-      att_wpn_knowledge = weapon_attack_skill(att, att_wpn_type);
-      if (is_in(OBJ_VALUE(att_wpn, 1), -25, 25))
+      if (pStat)
       {
-         tmp = OBJ_VALUE(att_wpn, 1) + OBJ_VALUE(att_wpn, 2);
-         att_bonus += tmp;
+         sprintf(buf, "Def not fighting att:          %4d<br/>", -25);
+         pStat->append(buf);
       }
    }
-   else
-   {
-      att_wpn_type = CHAR_ATTACK_TYPE(att);
-      att_wpn_knowledge = weapon_attack_skill(att, att_wpn_type);
-   }
 
-   def_wpn_knowledge = weapon_defense_skill(def, att_wpn_type);
-   sprintf(buf, "Weapon skill        :  %4d   %4d<br/>", att_wpn_knowledge, def_wpn_knowledge);
-   str.append(buf);
-
-   sprintf(buf, "Att Weapon mat+mag  :  %4d       <br/>", tmp);
-   str.append(buf);
-
-   tmp = 0;
-   if (char_dual_wield(att))
-   {
-      int dual_skill;
-
-      if (IS_PC(att))
-         dual_skill = PC_SKI_SKILL(att, SKI_DUAL_WIELD);
-      else
-         dual_skill = CHAR_DEX(att);
-
-      if (primary)
-      {
-         tmp = -MAX(0, 25 - (dual_skill / 4));
-         att_bonus += tmp;
-      }
-      else
-      {
-         tmp = -MAX(0, 50 - (dual_skill / 4));
-         att_bonus += tmp;
-      }
-   }
-   sprintf(buf, "Dual wield          :  %4d       <br/>", tmp);
-   str.append(buf);
-
-   tmp = 0;
-   if (CHAR_FIGHTING(def) != att)
-   {
-      tmp = -25;
-      def_bonus += tmp;
-   }
-
-   sprintf(buf, "Fighting attacker?  :  %4d       <br/>", tmp);
-   str.append(buf);
 
    /* If attacker can't see the defender, then the defender has a   */
    /* much better effective dexterity (since attacker can't see him */
    if (!CHAR_CAN_SEE(att, def))
    {
       def_bonus += 25;
-      sprintf(buf, "Attacker cant see    :       %4d  <br/>", 25);
-      str.append(buf);
+      if (pStat)
+      {
+         sprintf(buf, "Attacker cant see    :       %4d<br/>", 25);
+         pStat->append(buf);
+      }
    }
 
    if (!CHAR_CAN_SEE(def, att))
    {
-      sprintf(buf, "Defender cant see    :  %4d  <br/>", 25);
-      str.append(buf);
       att_bonus += 25;
+      if (pStat)
+      {
+         sprintf(buf, "Defender cant see    :  %4d<br/>", 25);
+         pStat->append(buf);
+      }
    }
 
    /* Slaying Weapons */
    if (att_wpn && OBJ_VALUE(att_wpn, 3) == CHAR_RACE(def))
    {
-      sprintf(buf, "Slaying              :  %4d  <br/>", 50);
-      str.append(buf);
       att_bonus += 50;
+
+      if (pStat)
+      {
+         sprintf(buf, "Slaying              :  %4d  <br/>", 50);
+         pStat->append(buf);
+      }
    }
 
    if ((UNIT_IS_GOOD(att) && affected_by_spell(def, ID_PROT_GOOD)) ||
        (UNIT_IS_EVIL(att) && affected_by_spell(def, ID_PROT_EVIL)))
    {
-      sprintf(buf, "Evil good protect     :       %4d  <br/>", 20);
-      str.append(buf);
       def_bonus += 20;
+
+      if (pStat)
+      {
+         sprintf(buf, "Evil good protect     :       %4d  <br/>", 20);
+         pStat->append(buf);
+      }
    }
 
    if ((def_armour = equipment_type(def, hit_loc, ITEM_ARMOR)))
@@ -633,8 +554,12 @@ void stat_melee_bonus(string &str,
       {
          tmp = OBJ_VALUE(def_armour, 1) + OBJ_VALUE(def_armour, 2);
          def_bonus += tmp;
-         sprintf(buf, "Armor bonus         :       %4d  <br/>", tmp);
-         str.append(buf);
+
+         if (pStat)
+         {
+            sprintf(buf, "Armor bonus         :       %4d  <br/>", tmp);
+            pStat->append(buf);
+         }
       }
       else
          slog(LOG_ALL, 0, "Illegal armour bonus.");
@@ -654,52 +579,62 @@ void stat_melee_bonus(string &str,
    if (pDef_armour)
       *pDef_armour = def_armour;
 
-   /*tmp = kludge_bonus(CHAR_LEVEL(att), att_wpn_knowledge);
-   att_wpn_knowledge += tmp;
-   tmp2 = kludge_bonus(CHAR_LEVEL(def), def_wpn_knowledge) / 2;
-   def_wpn_knowledge += tmp2;
-   sprintf(buf, "Kludge bonus        :  %4d   %4d<br/>", tmp, tmp2);
-   str.append(buf);*/
-
+   int def_dex = effective_dex(def);
    int att_abil = weapon_attack_ability(att, att_wpn_type);
-   sprintf(buf, "Abil for weapon     :  %4d<br/>", att_abil);
-   str.append(buf);
-
-   def_dex = effective_dex(def);
-   sprintf(buf, "Abil DEX DB         :         %4d<br/>", def_dex);
-   str.append(buf);
-
+   if (pStat)
+   {
+      sprintf(buf, "Abil wpn vs def DB  :  %4d    %4d<br/>", att_abil, def_dex);
+      pStat->append(buf);
+      sprintf(buf, "Wpn knowledge vs 1/3:  %4d    %4d<br/>", att_wpn_knowledge, def_wpn_knowledge/3);
+      pStat->append(buf);
+      sprintf(buf, "Att bonus vs Def    :  %4d    %4d<br/>", att_bonus, def_bonus);
+      pStat->append(buf);
+      sprintf(buf, "Level vs level      :  %4d    %4d<br/>", CHAR_LEVEL(att), CHAR_LEVEL(def));
+      pStat->append(buf);
+   }
 
    if (CHAR_AWAKE(def))
    {
-      hm = 2 * (att_abil - def_dex + att_wpn_knowledge - def_wpn_knowledge/3 + att_bonus - def_bonus);
-      sprintf(buf, "Result              :  %4d<br/>", hm);
-      str.append(buf);
+      hm = 2* ((att_abil - def_dex) + (att_wpn_knowledge - def_wpn_knowledge/3) + 
+                att_bonus - def_bonus + CHAR_LEVEL(att) - CHAR_LEVEL(def));
+      if (pStat)
+      {
+         sprintf(buf, "Result              :  %4d<br/>", hm);
+         pStat->append(buf);
+      }
    }
    else
    {
-      hm = 2 * (att_abil + att_wpn_knowledge + att_bonus);
-      sprintf(buf, "Result (sleep def)  :  %4d<br/>", hm);
-      str.append(buf);
+      hm = 2* (att_abil + att_wpn_knowledge + att_bonus + CHAR_LEVEL(att) - CHAR_LEVEL(def));
+      if (pStat)
+      {
+         sprintf(buf, "Result (sleep def)  :  %4d<br/>", hm);
+         pStat->append(buf);
+      }
    }
-   sprintf(buf, "Result = 2*(att_abil-def_abil + att_wpn - def_wpn + att_bonus - def_bonus) + roll<br/>");
-   str.append(buf);
 
-   str.append("<br/>");
+   if (pStat)
+   {
+      pStat->append("Result = 2*((att_abil-def_dex) + (att_wpn - def_wpn/3) + (att_bonus - def_bonus) + (att level - def level)) + roll<br/>");
+      pStat->append("<br/>");
 
-   int dam5  = weapon_damage(5 + hm, att_wpn_type, def_armour_type);
-   int dam50 = weapon_damage(50 + hm, att_wpn_type, def_armour_type);
-   int dam95 = weapon_damage(95 + hm, att_wpn_type, def_armour_type);
+      int dam5  = weapon_damage(5 + hm, att_wpn_type, def_armour_type);
+      int dam50 = weapon_damage(50 + hm, att_wpn_type, def_armour_type);
+      int dam95 = weapon_damage(95 + hm, att_wpn_type, def_armour_type);
 
-   sprintf(buf, "Weapon dmg (5/50/95) : %4d %4d %4d<br/>", dam5, dam50, dam95);
-   str.append(buf);
+      sprintf(buf, "Weapon dmg (5/50/95) : %4d %4d %4d<br/>", dam5, dam50, dam95);
+      pStat->append(buf);
 
-   sprintf(buf, "Rounds to kill def = %d<br/>", UNIT_MAX_HIT(def)/MAX(1, (dam5 + dam50 + dam95) / 3));
-   str.append(buf);
+      sprintf(buf, "Rounds to kill def = %d<br/>", UNIT_MAX_HIT(def)/MAX(1, (dam5 + dam50 + dam95) / 3));
+      pStat->append(buf);
 
+      pStat->append("</pre>");
+   }
 
-   str.append("</pre>");
+   return MAX(-50, hm);
 }
+
+
 
 int base_melee(class unit_data *att, class unit_data *def, int hit_loc)
 {
@@ -725,7 +660,7 @@ int base_melee(class unit_data *att, class unit_data *def, int hit_loc)
 /* danger involved in combat (i.e. how fast would you die in worst case)    */
 /* Returns number of rounds it takes att to kill def                        */
 
-int base_consider(class unit_data *att, class unit_data *def)
+int base_consider(class unit_data *att, class unit_data *def, string *pStr)
 {
    int ocp, bonus;
    int att_wpn_type, def_arm_type;
@@ -737,15 +672,18 @@ int base_consider(class unit_data *att, class unit_data *def)
 
    att_wpn_type = WPN_ROOT;
 
-   bonus = melee_bonus(att, def, WEAR_BODY,
-                       &att_wpn_type, NULL, &def_arm_type, NULL);
+   bonus = melee_bonus(att, def, WEAR_BODY, &att_wpn_type, NULL, &def_arm_type, NULL, 1, pStr);
+   bonus -= 50; // Compensate for the 2*-25 for def not fighting attacker
 
    CHAR_POS(def) = ocp;
 
-   dam = weapon_damage(50 + bonus, att_wpn_type, def_arm_type);
+   dam  = weapon_damage( 5 + bonus, att_wpn_type, def_arm_type);
+   dam += weapon_damage(50 + bonus, att_wpn_type, def_arm_type);
+   dam += weapon_damage(95 + bonus, att_wpn_type, def_arm_type);
+   dam /= 3;
 
    if (dam <= 0)
-      return 10000;
+      return 100;
    else
       return UNIT_MAX_HIT(def) / dam; /* Rounds to die.... */
 }
@@ -755,6 +693,7 @@ void do_consider(class unit_data *ch, char *arg, const struct command_info *cmd)
    class unit_data *vict;
    int rtd;
    char *oarg = arg;
+   string str;
 
    if (IS_PC(ch) && PC_SKI_SKILL(ch, SKI_CONSIDER) == 0)
    {
@@ -799,14 +738,17 @@ void do_consider(class unit_data *ch, char *arg, const struct command_info *cmd)
           A_SOMEONE, ch, cActParameter(), vict, TO_CHAR);
    }
 
-   rtd = base_consider(vict, ch);
+   rtd = base_consider(vict, ch, &str);
 
    if (IS_IMMORTAL(ch))
    {
       act("Rounds to be killed by $3n = $2d.", A_SOMEONE, ch, &rtd, vict, TO_CHAR);
+      send_to_char(str.c_str(), ch);
 
-      rtd = base_consider(ch, vict);
+      str = "";
+      rtd = base_consider(ch, vict, &str);
       act("Rounds to kill $3n = $2d.", A_SOMEONE, ch, &rtd, vict, TO_CHAR);
+      send_to_char(str.c_str(), ch);
       return;
    }
